// apps/sales-copilot/src/generated/services/Crf5c_activity1sService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { Crf5c_activity1sBase, Crf5c_activity1s } from '../models/Crf5c_activity1sModel';
import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Crf5c_activity1sService {
  private static readonly dataSourceName = 'crf5c_activity1s';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Crf5c_activity1sBase, 'crf5c_activity1id'>): Promise<IOperationResult<Crf5c_activity1s>> {
    const result = await Crf5c_activity1sService.client.createRecordAsync<Omit<Crf5c_activity1sBase, 'crf5c_activity1id'>, Crf5c_activity1s>(
      Crf5c_activity1sService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Crf5c_activity1sBase, 'crf5c_activity1id'>>): Promise<IOperationResult<Crf5c_activity1s>> {
    const result = await Crf5c_activity1sService.client.updateRecordAsync<Partial<Omit<Crf5c_activity1sBase, 'crf5c_activity1id'>>, Crf5c_activity1s>(
      Crf5c_activity1sService.dataSourceName,
      id.toString(),
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Crf5c_activity1sService.client.deleteRecordAsync(
      Crf5c_activity1sService.dataSourceName,
      id.toString());
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Crf5c_activity1s>> {
    const result = await Crf5c_activity1sService.client.retrieveRecordAsync<Crf5c_activity1s>(
      Crf5c_activity1sService.dataSourceName,
      id.toString(),
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Crf5c_activity1s[]>> {
    const result = await Crf5c_activity1sService.client.retrieveMultipleRecordsAsync<Crf5c_activity1s>(
      Crf5c_activity1sService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Crf5c_activity1s> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Crf5c_activity1sService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Crf5c_activity1sService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }
}