// apps/sales-copilot/src/generated/models/Crf5c_activity1sModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crf5c_activity1scrf5c_draftstatus = {
  995340000: 'draft',
  995340001: 'confirmed',
  995340002: 'completed',
  995340003: 'cancelled'
} as const;
export type Crf5c_activity1scrf5c_draftstatus = keyof typeof Crf5c_activity1scrf5c_draftstatus;
export const Crf5c_activity1scrf5c_outcome = {
  995340000: '__',
  995340001: '__',
  995340002: '____',
  995340003: '_____',
  995340004: '___'
} as const;
export type Crf5c_activity1scrf5c_outcome = keyof typeof Crf5c_activity1scrf5c_outcome;
export const Crf5c_activity1scrf5c_type = {
  995340000: 'visit',
  995340001: 'call',
  995340002: 'meeting',
  995340003: 'email',
  995340004: 'other'
} as const;
export type Crf5c_activity1scrf5c_type = keyof typeof Crf5c_activity1scrf5c_type;
export const Crf5c_activity1sstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crf5c_activity1sstatecode = keyof typeof Crf5c_activity1sstatecode;
export const Crf5c_activity1sstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crf5c_activity1sstatuscode = keyof typeof Crf5c_activity1sstatuscode;

export interface Crf5c_activity1sBase {
  "biz_Contact@odata.bind"?: string;
  "crf5c_Account@odata.bind"?: string;
  crf5c_activity1id: string;
  crf5c_createdon?: string;
  crf5c_draftstatus: Crf5c_activity1scrf5c_draftstatus;
  crf5c_notes?: string;
  "crf5c_Opportunity@odata.bind"?: string;
  crf5c_outcome?: Crf5c_activity1scrf5c_outcome;
  crf5c_ownerid: string;
  crf5c_scheduleddate: string;
  crf5c_title: string;
  crf5c_type: Crf5c_activity1scrf5c_type;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crf5c_activity1sstatecode;
  statuscode?: Crf5c_activity1sstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Crf5c_activity1s extends Crf5c_activity1sBase {
  biz_contactname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  crf5c_accountname?: string;
  crf5c_draftstatusname?: string;
  crf5c_opportunityname?: string;
  crf5c_outcomename?: string;
  crf5c_typename?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  biz_contact?: object;
  _biz_contact_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  crf5c_account?: object;
  _crf5c_account_value?: string;
  crf5c_opportunity?: object;
  _crf5c_opportunity_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}